fix: close the rest of the Linux/clang warning sweep
Covers the remaining categories from the full warning sweep:
format-security (ImGui::Text/TextWrapped/TextDisabled/TextColored/
BulletText/Util::Text::* helpers with a runtime, no-arg format string
-> TextUnformatted or an explicit "%s" literal), missing-field-initializers
(brace-init structs/aggregates missing a trailing field), sign-compare
(size_t loop vars for pure indexing, static_cast<int> where a callee/
sentinel needs signed), unused-function/variable/const-variable/
lambda-capture/private-field (delete genuinely dead code), missing-braces
(subobject brace wrapping), logical-op-parentheses, and a dependent-name
.template get<>() fix.

Real bugs found and fixed along the way:
- LightPicker.cpp/UnifiedWater.cpp: !file->fileName tested a fixed-size
  array's address, always true; fixed to test emptiness.
- BSShaderHooks.cpp: the Lighting-technique replacement path tracked
  found/success/failed counts but never logged them, unlike its sibling
  path a few lines down; added the matching summary log.
- ShaderCache.cpp: two ShaderClass switches were missing Total; two
  BSShader::Type switches were missing None/Total.

Verified via the WSL clang-cl-20 cross-compile toolchain.

+++ src/FrameAnnotations.cpp
@@ -1091,7 +1091,7 @@ namespace FrameAnnotations
 		auto renderer = globals::game::renderer;
 
 		for (size_t renderTargetIndex = 0;
-			renderTargetIndex < Util::GetRenderTargetCount(); ++renderTargetIndex) {
+			renderTargetIndex < static_cast<size_t>(Util::GetRenderTargetCount()); ++renderTargetIndex) {
 			const auto renderTargetName = magic_enum::enum_name(
 				static_cast<RE::RENDER_TARGETS::RENDER_TARGET>(renderTargetIndex));
 			if (auto texture = renderer->GetRuntimeData().renderTargets[renderTargetIndex].texture) {
@@ -1112,7 +1112,7 @@ namespace FrameAnnotations
 		}
 
 		for (size_t renderTargetIndex = 0;
-			renderTargetIndex < Util::GetDepthStencilCount();
+			renderTargetIndex < static_cast<size_t>(Util::GetDepthStencilCount());
 			++renderTargetIndex) {
 			const auto renderTargetName = magic_enum::enum_name(
 				static_cast<RE::RENDER_TARGETS_DEPTHSTENCIL::RENDER_TARGET_DEPTHSTENCIL>(
